Parser.py
Added Twilio and support for text notifications. Note: it is critical for you to use your own values for the Twilio stuff. I changed it to be different from mine because I don't want to get spammed (also you wouldn't see it because I would be getting the texts not you).